requirements.txt

opencv-python
picamera2
RPi.GPIO
numpy